requirements.txt
scapy-python3
python-nmap
netifaces